of: of_batterydata: Print battery resistor value on probe
We need it to know all the battery resistor IDs for all Sony
devices, hence print the value once probed.
IF/when people will have bad battery behavior, this will mean
that they've got a different batt, hence we will be able to
implement its data.

Revert this commit when job is done.